chore: UserManager에 뱃지업데이트 함수 추가
